+++ CHANGELOG.md
@@ -16,11 +16,12 @@
   To use the new collection replace `allowInsecureProtocols : true` with `allowedSchemes : ["https", "http", "wss", "ws"]`.
   You can remove `wss` and `ws` if you have no WebSocket use.
 
-* **Breaking** `ProxiedSsrfDelgatingHandler` now takes a new options class, `ProxiedSsrfOptions`, instead of `SsrfOptions` to allow for proxy specific configuration. The new options class inherits from `SsrfOptions` so all existing configuration options are still available,
-  and the `Proxy` property has been added
+ * **Breaking** `ProxiedSsrfDelegatingHandler` now takes a new options class, `ProxiedSsrfOptions`, instead of `SsrfOptions` to allow for proxy specific configuration. The new options class inherits from `SsrfOptions`
+   so all existing configuration options are still available, and the `Proxy` property has been added
   * `Proxy` - an instance of `WebProxy` that will be used for the handler.
 
-* **Breaking** `ProxiedSsrfDelgatingHandler` constructor now takes a `WebProxy` instance rather than an `IWebProxy` instance to allow the automatic safe listing of the proxy address.
+* **Breaking** `ProxiedSsrfDelegatingHandler` constructor now takes a `WebProxy` instance rather than an `IWebProxy` instance to allow the automatic safe listing of the proxy address.
+of the proxy address.
 
 * Update OTEL dependencies to address [CVE-2026-40894 - OpenTelemetry dotnet: Excessive memory allocation when parsing OpenTelemetry propagation headers](https://github.com/advisories/GHSA-g94r-2vxg-569j)